+++ app.js
@@ -19,6 +19,7 @@ var contact = require('./routes/contact');
 var dashboard = require('./routes/dashboard');
 var adminview = require('./routes/adminview');
 var ordersplaced = require('./routes/ordersplaced');
+var orderdetails = require('./routes/orderdetails');
 
 var app = express();
 
@@ -47,6 +48,7 @@ app.use('/users/receivers', receivers);
 app.use('/users/dashboard', dashboard);
 app.use('/users/ordersplaced', ordersplaced);
 app.use('/users/delivers_select', delivers_select);
+app.use('/users/orderdetails', orderdetails);
 app.use('/about', about);
 app.use('/contact', contact);
 app.use('/adminview', adminview);
