+++ views/orderdetails.hbs
@@ -1,0 +1,37 @@
+<h3 id="orderdetails_head">Order Details</h3>
+<p>The content displayed below is the details of your orders...</p>
+
+{{#with order}}
+<div  id="orderdetails_table">	
+	<div class="row">
+		<div class="col-md-12">	
+			<h4 id="table_tittle">Order ID #: {{_id}}</h4>
+		</div>
+	</div>
+	<div class="row">
+		<div class="col-md-12">
+			<h4 id="table_tittle">Delivery Address: </h4>
+			<div id="orderdetails">
+				<p>{{receiversinfo.adds}}</p> 
+				<p>ZIP {{receiversinfo.zip}}</p>
+			</div>
+		</div>
+	</div>
+	<div class="row" id="orderdetails_table_item">
+		<div class="col-md-6">
+			<h4 id="table_tittle">Item Bought: </h4>
+			<div id="orderdetails">
+				<p>{{iteminfo.item}}</p>
+				<h5>Price: ${{iteminfo.price}}</h5> 
+			</div>
+		</div>
+		<div class="col-md-6">
+			<h4 id="table_tittle">Store Location: </h4>
+			<div id="orderdetails">
+				<p>{{iteminfo.store_location}}</p>
+				<p>ZIP {{iteminfo.store_zip}}</p>
+			</div>
+		</div>
+	</div>
+</div>
+{{/with}}
