+++ views/orders_placed.hbs
@@ -3,7 +3,7 @@
     <h2>List of placed orders</h2>
     <div id="orderlist">
       {{#orders}}
-      <a href="/users/orderdetail?objectId={{_id}}" class="list-group-item" id="orderlist_item">
+      <a href="/users/orderdetails?objectId={{_id}}" class="list-group-item" id="orderlist_item">
         <h4 class="list-group-item-heading">
           {{iteminfo.item}} placed at {{created_at}}</h4>
         <p class="list-group-item-text">
